+++ internal/provider/resources/resource_datapower_config_sequence.go
@@ -148,7 +148,6 @@ func (r *ConfigSequenceResource) Schema(ctx context.Context, req resource.Schema
 				},
 				Default: int64default.StaticInt64(100),
 			},
-			"capabilities":       models.GetDmConfigSequenceCapabilitiesResourceSchema("Capabilities", "", "", false),
 			"dependency_actions": actions.ActionsSchema,
 		},
 	}